AP-8423: Allow port object to hold directories
The DLTensorFlowNetworkPortObject used
DLPortObject.copyFileToFileStore() to save the model in the file store.
This method doesn't work for directories which are allowed for
TensorFlow models.
A TensorFlow network now needs to know how to store its relevant files
in a file store since this could be different for different networks.