AP-8939: Add DialogComponent for arbitrary objects
The DialogComponent is called DialogComponentObjectSelection and can
show a JComboBox for different objects. The caller must provide:
- A function to print the object as a String
- A function that saves an object in an SettingsModel
- A function that gets an object from an SettingsModel

This new DialogComponent is used to select arrays of tags.

NOTE: There is already a similar DialogComponent in org.knime.dl. But
this one doesn't use SettingsModels. We should find a way to merge them.