AP-8425: Save whole name of tensor in spec
The name of a tensor is constructed like '<op_name>_id:<output_id>'.
The fetch of the tensor requires '<op_name>_id' as a String and
<output_id> as an integer.

Previously only '<op_name>_id' was saved and <output_id> was assumed
to be 0 which is fine in most cases. Now everything is saved and parsed
directly before the graph is executed.

(Note that '<op_name>_id' could also be a name set by the user)