Fix for pandas<0.21.0: Not use scalars
For pandas<0.21.0 it doesn't work to create a DataFrame with scalars and
without using the index parameter but this was done for the tags which
can be lists or single strings.
Note, that a Sequence can also be a RepeatedScalarContainer but pandas
can't handle this (in lower versions) so they are written into a new
list.