AP-10534: Adapt converters to interface move
We keep them in the tensorflow plugin as long as we can't deprecate
them. Once this is possible we should move them to dl-core.
A consequence is that users can't use Strings as inputs if they have
only installed the Keras integration but this is not a real limitation
since Keras can't do anything useful with string tensors anyway (at
least not yet).

AP-10534 (DL Framework: Support String inputs)