AP-11382: Check if knime protocol file is local
Local and non local files are handled differently because of performance
reasons. The check if a file is local or non-local was only done by
checking the protocol of the URL. But knime URLs can be both, local or
non-local.
This commit adds a check for handling the non-local knime files.