AP-14657: Load TF2 model uncompiled if compiled fails
If a custom loss function was used to train the model loading it
compiled will fail because the loss function is not known and not saved
as part of the SavedModel. Loading it uncompiled works and is sufficient
for running the model.

AP-14657 (Tensorflow 2 Executer does not run for individually defined loss functions)