AP-16042: Fix settings loading in TF2 Writer Dialog
The TF2WriterNodeDialog did override the wrong method to load settings.
It did override the method for DataTableSpec but the node has a port
object input and therefore needed to override the method for
PortObjectSpec. This method was not called which caused a NPE in the
PythonCommandFlowVariableModel on dialog open.

AP-16042 (TensorFlow2 Network Writer node cannot be configured)